gui separated from WorldManager
