+++ DwarfCorp/DwarfCorpCore/DwarfGame.cs
@@ -105,7 +105,7 @@ namespace DwarfCorp
             StateManager.States["CompanyMakerState"] = new CompanyMakerState(this, StateManager);
             StateManager.States["WorldLoaderState"] = new WorldLoaderState(this, StateManager);
             StateManager.States["GameLoaderState"] = new GameLoaderState(this, StateManager);
-            StateManager.States["LoseState"] = new LoseState(this, StateManager, playState.World);
+            StateManager.States["LoseState"] = new LoseState(this, StateManager, playState);
 
             if(GameSettings.Default.DisplayIntro)
             {
