+++ DwarfCorp/DwarfCorpCore/Scripting/Player/Tools/BuildTool.cs
@@ -65,7 +65,7 @@ namespace DwarfCorp
             }
             int w = 600;
             int h = 350;
-            BuildPanel = new BuildMenu(PlayState.GUI, PlayState.GUI.RootComponent, Player, BuildType)
+            BuildPanel = new BuildMenu(WorldManager.GUI, WorldManager.GUI.RootComponent, Player, BuildType)
             {
                 LocalBounds = new Rectangle(PlayState.Game.GraphicsDevice.Viewport.Width/2 - w/2, PlayState.Game.GraphicsDevice.Viewport.Height/2 - h/2, w, h),
                 IsVisible = true,
@@ -85,7 +85,7 @@ namespace DwarfCorp
             if (Player.IsCameraRotationModeActive())
             {
                 Player.VoxSelector.Enabled = false;
-                PlayState.GUI.IsMouseVisible = false;
+                WorldManager.GUI.IsMouseVisible = false;
                 Player.BodySelector.Enabled = false;
                 return;
             }
@@ -97,9 +97,9 @@ namespace DwarfCorp
             {
                 Player.VoxSelector.Enabled = true;
                 Player.BodySelector.Enabled = false;
-                PlayState.GUI.IsMouseVisible = true;
+                WorldManager.GUI.IsMouseVisible = true;
 
-                PlayState.GUI.MouseMode = PlayState.GUI.IsMouseOver()
+                WorldManager.GUI.MouseMode = WorldManager.GUI.IsMouseOver()
                     ? GUISkin.MousePointer.Pointer
                     : GUISkin.MousePointer.Build;
             }
@@ -107,9 +107,9 @@ namespace DwarfCorp
             {
                 Player.VoxSelector.Enabled = false;
                 Player.BodySelector.Enabled = false;
-                PlayState.GUI.IsMouseVisible = true;
+                WorldManager.GUI.IsMouseVisible = true;
 
-                PlayState.GUI.MouseMode = PlayState.GUI.IsMouseOver()
+                WorldManager.GUI.MouseMode = WorldManager.GUI.IsMouseOver()
                     ? GUISkin.MousePointer.Pointer
                     : GUISkin.MousePointer.Cook;
             }
