+++ DwarfCorp/DwarfCorpCore/Scripting/Player/Tools/ChopTool.cs
@@ -72,22 +72,22 @@ namespace DwarfCorp
             {
                 Player.VoxSelector.Enabled = false;
                 Player.BodySelector.Enabled = false;
-                PlayState.GUI.IsMouseVisible = false;
+                WorldManager.GUI.IsMouseVisible = false;
                 return;
             }
 
             Player.VoxSelector.Enabled = false;
             Player.BodySelector.Enabled = true;
             Player.BodySelector.AllowRightClickSelection = true;
-            PlayState.GUI.IsMouseVisible = true;
+            WorldManager.GUI.IsMouseVisible = true;
 
-            if (PlayState.GUI.IsMouseOver())
+            if (WorldManager.GUI.IsMouseOver())
             {
-                PlayState.GUI.MouseMode = GUISkin.MousePointer.Pointer;
+                WorldManager.GUI.MouseMode = GUISkin.MousePointer.Pointer;
             }
             else
             {
-                PlayState.GUI.MouseMode = GUISkin.MousePointer.Chop;
+                WorldManager.GUI.MouseMode = GUISkin.MousePointer.Chop;
             }
         }
