+++ DwarfCorp/DwarfCorpCore/Scripting/Player/Tools/FarmTool.cs
@@ -131,12 +131,12 @@ namespace DwarfCorp
                         {
                             if (!voxel.Type.IsSoil)
                             {
-                                PlayState.GUI.ToolTipManager.Popup("Can only till soil!");
+                                WorldManager.GUI.ToolTipManager.Popup("Can only till soil!");
                                 continue;
                             }
                             if (voxel.TypeName == "TilledSoil")
                             {
-                                PlayState.GUI.ToolTipManager.Popup("Soil already tilled!");
+                                WorldManager.GUI.ToolTipManager.Popup("Soil already tilled!");
                                 continue;
                             }
                             if (!HasTile(voxel))
@@ -179,12 +179,12 @@ namespace DwarfCorp
 
                         if (currentAmount == 0)
                         {
-                            PlayState.GUI.ToolTipManager.Popup("Not enough " + PlantType + " in stocks!");
+                            WorldManager.GUI.ToolTipManager.Popup("Not enough " + PlantType + " in stocks!");
                             break;
                         }
                         if (voxel.TypeName != "TilledSoil")
                         {
-                            PlayState.GUI.ToolTipManager.Popup("Can only plant on tilled soil!");
+                            WorldManager.GUI.ToolTipManager.Popup("Can only plant on tilled soil!");
                             continue;
                         }
 
@@ -192,7 +192,7 @@ namespace DwarfCorp
                         {
                             if (voxel.SunColor == 0)
                             {
-                                PlayState.GUI.ToolTipManager.Popup("Can only plant " + PlantType + " above ground.");
+                                WorldManager.GUI.ToolTipManager.Popup("Can only plant " + PlantType + " above ground.");
                                 continue;
                             }
                         }
@@ -202,7 +202,7 @@ namespace DwarfCorp
                         {
                             if (voxel.SunColor > 0)
                             {
-                                PlayState.GUI.ToolTipManager.Popup("Can only plant " + PlantType + " below ground.");
+                                WorldManager.GUI.ToolTipManager.Popup("Can only plant " + PlantType + " below ground.");
                                 continue;
                             }
                         }
@@ -216,7 +216,7 @@ namespace DwarfCorp
                         }
                         else
                         {
-                            PlayState.GUI.ToolTipManager.Popup("Something is already planted here!");
+                            WorldManager.GUI.ToolTipManager.Popup("Something is already planted here!");
                             continue;
                         }
                     }
@@ -265,7 +265,7 @@ namespace DwarfCorp
             }
             int w = 600;
             int h = 350;
-            FarmPanel = new FarmingPanel(PlayState.GUI, PlayState.GUI.RootComponent)
+            FarmPanel = new FarmingPanel(WorldManager.GUI, WorldManager.GUI.RootComponent)
             {
                 LocalBounds = new Rectangle(PlayState.Game.GraphicsDevice.Viewport.Width / 2 - w / 2, PlayState.Game.GraphicsDevice.Viewport.Height / 2 - h / 2, w, h),
                 IsVisible = true,
@@ -279,13 +279,13 @@ namespace DwarfCorp
 
         void FarmPanel_OnTill()
         {
-            PlayState.GUI.ToolTipManager.Popup("Click and drag to till soil.");
+            WorldManager.GUI.ToolTipManager.Popup("Click and drag to till soil.");
             Mode = FarmMode.Tilling;
         }
 
         void FarmPanel_OnPlant(string plantType, string resource)
         {
-            PlayState.GUI.ToolTipManager.Popup("Click and drag to plant " + plantType + ".");
+            WorldManager.GUI.ToolTipManager.Popup("Click and drag to plant " + plantType + ".");
             Mode = FarmMode.Planting;
             PlantType = plantType;
             RequiredResources = new List<ResourceAmount>() {new ResourceAmount(resource)};
@@ -293,7 +293,7 @@ namespace DwarfCorp
 
         void FarmPanel_OnHarvest()
         {
-            PlayState.GUI.ToolTipManager.Popup("Click and drag to harvest.");
+            WorldManager.GUI.ToolTipManager.Popup("Click and drag to harvest.");
             Mode = FarmMode.Harvesting;
         }
 
@@ -308,7 +308,7 @@ namespace DwarfCorp
             if (Player.IsCameraRotationModeActive())
             {
                 Player.VoxSelector.Enabled = false;
-                PlayState.GUI.IsMouseVisible = false;
+                WorldManager.GUI.IsMouseVisible = false;
                 Player.BodySelector.Enabled = false;
                 return;
             }
@@ -330,9 +330,9 @@ namespace DwarfCorp
                     Player.BodySelector.Enabled = true;
                     break;
             }
-            PlayState.GUI.IsMouseVisible = true;
+            WorldManager.GUI.IsMouseVisible = true;
 
-            PlayState.GUI.MouseMode = PlayState.GUI.IsMouseOver() ? GUISkin.MousePointer.Pointer : GUISkin.MousePointer.Farm;
+            WorldManager.GUI.MouseMode = WorldManager.GUI.IsMouseOver() ? GUISkin.MousePointer.Pointer : GUISkin.MousePointer.Farm;
         }
 
         public override void Render(DwarfGame game, GraphicsDevice graphics, DwarfTime time)
