+++ DwarfCorp/DwarfCorpCore/Scripting/Rooms/RoomTypes/WheatFarm.cs
@@ -94,7 +94,7 @@ namespace DwarfCorp
         public override void OnBuilt()
         {
            
-            Button farmButton = new Button(PlayState.GUI, PlayState.GUI.RootComponent, "Farm", PlayState.GUI.DefaultFont,
+            Button farmButton = new Button(WorldManager.GUI, WorldManager.GUI.RootComponent, "Farm", WorldManager.GUI.DefaultFont,
                 Button.ButtonMode.ImageButton, new NamedImageFrame(ContentPaths.GUI.icons, 32, 5, 1))
             {
                 LocalBounds = new Rectangle(0, 0, 32, 32),
